+++ 2_simulation/0_parameter/transmittance.py
@@ -113,7 +113,7 @@ def run(parameter):
         images *= np.exp(-mu * THICKNESS * 1e-3 * simpli.voxel_size)
 
         simpli.noise_model = lambda x: np.round(
-            np.random.normal(x, np.sqrt(gain * x))).astype(np.uint16)
+            np.random.normal(x, np.sqrt(gain * x))).astype(np.float32)
 
         _, images_ = simpli.apply_optic(images)
         t, d, r = simpli.apply_epa(images_)
